HTTP 1.1: Forward OPTIONS requests properly
Bump the max_forwards type to 64-bit to cope with wider range of values.

NP: still no internal handling of OPTIONS requests, but this will
     forward them according to RFC 2616 requirements.

+++ src/HttpRequest.h
@@ -96,7 +96,7 @@ public:
 
     int imslen;
 
-    int max_forwards;
+    int64_t max_forwards;
 
     /* these in_addr's could probably be sockaddr_in's */
